Adicionado insert de Receita em ReceitaDAO

+++ PED12/Sistema_finaceiro_PI/src/DAO/CartaoDebitoDAO.java
@@ -5,7 +5,6 @@
  */
 package DAO;
 
-import Model.CartaoCredito;
 import Model.CartaoDebito;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -18,6 +17,7 @@ import javax.swing.JOptionPane;
  */
 public class CartaoDebitoDAO {
     
+    
     private Connection conexao = null;
 
     public CartaoDebitoDAO() {
@@ -26,23 +26,23 @@ public class CartaoDebitoDAO {
     
     }
     
-    public boolean CadastrarCartaoDebito(CartaoDebito cartao_db) throws SQLException{
-        
-        PreparedStatement pst = null;
-        
-        String insert = "insert into cartao_debito (n_cartao_debito,valor_atual, bandeira, conta_id_conta) values(?,?, ?, ?)";
-        
-        pst = conexao.prepareStatement(insert);
-        
+   public boolean CadastrarCartaoDebito(CartaoDebito cartao_db) throws SQLException {
+       
+       PreparedStatement pst = null;
+       
+       String insert = "insert into cartao_debito (n_cartao_debito,valor_atual, bandeira, conta_id_conta) values(?,?, ?, ?)";
+
+       pst = conexao.prepareStatement(insert);
+       
         try {
 
             pst.setLong(1, cartao_db.getN_cartao_debito());
             pst.setFloat(2, cartao_db.getValor_atual());
             pst.setString(3, cartao_db.getBandeira());
-            pst.setInt(4, cartao_db.getId_conta());
+            pst.setFloat(4, cartao_db.getId_conta());
             
             pst.executeUpdate();
-            
+
             return true;
 
         } catch (Exception e) {
@@ -51,12 +51,12 @@ public class CartaoDebitoDAO {
 
             return false;
 
-        }finally{
-            
+        } finally {
+
             pst.close();
-            
+
         }
-        
+
     }
     
 }
